requirements (4).txt
python-telegram-bot[ext]
google-generativeai
python-dotenv
Pillow
gradio
# Thêm các thư viện này để có thể giúp ổn định kết nối mạng
httpx[http2]
certifi